requirements.txt
django==1.8.6
django-cms==3.2.2
django-filer==1.1.1
